# apps/api/chats/router.py
"""Chat CRUD, messages, run creation (TRD §12). Ownership filters are
injected server-side on every query (CLAUDE.md non-negotiable)."""

from uuid import UUID

from fastapi import APIRouter, Request, UploadFile
from sqlalchemy import select
from sqlalchemy.ext.asyncio import AsyncSession

from auth.deps import CurrentUser
from chats.scope import (
    chat_starter_questions,
    dedupe,
    library_starter_questions,
    resolve_scope,
)
from db.models import (
    Chat,
    Chunk,
    Citation,
    Collection,
    Document,
    LlmProvider,
    Message,
    Model,
    ModelRole,
    Run,
    Section,
    User,
)
from db.session import SessionDep
from errors import AppError
from graph import runner
from graph.chat_title import DEFAULT_CHAT_TITLE, collapse_instant_title
from ingest.containers import get_chat_collection, get_or_create_chat_collection
from ingest.repository import list_container_documents
from ingest.upload import accept_upload, deny_read_only, document_out, read_upload
from quota.service import gate_and_reserve
from runtime import load_active_runtime
from schemas.chats import (
    ChatCreate,
    ChatOut,
    ChatPatch,
    CitationOut,
    MessageOut,
    RunCreateRequest,
    RunCreateResponse,
    chat_out,
    message_out,
)
from schemas.sources import DocumentOut, DocumentUploadOut

router = APIRouter(tags=["chats"])


class ChatNotFound(AppError):
    status_code = 404


class ModelNotAvailable(AppError):
    status_code = 422


async def _owned_chat(session: AsyncSession, user: User, chat_id: UUID) -> Chat:
    chat = (
        await session.execute(select(Chat).where(Chat.id == chat_id, Chat.user_id == user.id))
    ).scalar_one_or_none()
    if chat is None:
        raise ChatNotFound("chat_not_found", "Chat not found")
    return chat


async def _default_model_id(session: AsyncSession) -> str:
    role = (
        await session.execute(select(ModelRole).where(ModelRole.role == "generator"))
    ).scalar_one_or_none()
    if role is not None:
        return role.model_id
    first = (
        (await session.execute(select(Model).where(Model.enabled).order_by(Model.model_id)))
        .scalars()
        .first()
    )
    if first is None:
        raise AppError("no_models", "No models are configured", status_code=500)
    return first.model_id


@router.get("/chats", response_model=list[ChatOut])
async def list_chats(
    user: CurrentUser,
    session: SessionDep,
) -> list[ChatOut]:
    active_run = (
        select(Run.id)
        .join(Message, Run.message_id == Message.id)
        .where(Message.chat_id == Chat.id, Run.status == "running")
        .order_by(Run.created_at.desc())
        .limit(1)
        .scalar_subquery()
    )
    rows = (
        await session.execute(
            select(Chat, active_run).where(Chat.user_id == user.id).order_by(Chat.created_at.desc())
        )
    ).all()
    questions = await _starter_questions_by_chat(session, user, [chat for chat, _ in rows])
    return [
        chat_out(
            chat.id,
            chat.title,
            chat.pinned,
            chat.model_id,
            questions.get(chat.id, []),
            chat.created_at,
            active,
        )
        for chat, active in rows
    ]


async def _starter_questions_by_chat(
    session: AsyncSession, user: User, chats: list[Chat]
) -> dict[UUID, list[str]]:
    """One query for every chat's own questions; the Library is the fallback."""
    own_rows = await session.execute(
        select(Collection.chat_id, Collection.starter_questions).where(
            Collection.kind == "chat", Collection.owner_id == user.id
        )
    )
    own: dict[UUID, list[str]] = {}
    for chat_id, questions in own_rows.all():
        if chat_id is not None:
            own[chat_id] = dedupe(q for q in (questions or []))
    if any(not own.get(chat.id) for chat in chats):
        library = await library_starter_questions(session)
        for chat in chats:
            if not own.get(chat.id):
                own[chat.id] = library
    return own


@router.post("/chats", response_model=ChatOut, status_code=201)
async def create_chat(
    body: ChatCreate,
    user: CurrentUser,
    session: SessionDep,
) -> ChatOut:
    chat = Chat(
        user_id=user.id,
        title=body.title or "New chat",
        model_id=await _default_model_id(session),
    )
    session.add(chat)
    await session.flush()
    return chat_out(
        chat.id,
        chat.title,
        chat.pinned,
        chat.model_id,
        await chat_starter_questions(session, chat),
        chat.created_at,
        None,
    )


@router.get("/chats/{chat_id}", response_model=ChatOut)
async def get_chat(
    chat_id: UUID,
    user: CurrentUser,
    session: SessionDep,
) -> ChatOut:
    chat = await _owned_chat(session, user, chat_id)
    active_run = (
        await session.execute(
            select(Run.id)
            .join(Message, Run.message_id == Message.id)
            .where(Message.chat_id == chat.id, Run.status == "running")
            .order_by(Run.created_at.desc())
            .limit(1)
        )
    ).scalar_one_or_none()
    return chat_out(
        chat.id,
        chat.title,
        chat.pinned,
        chat.model_id,
        await chat_starter_questions(session, chat),
        chat.created_at,
        active_run,
    )


@router.patch("/chats/{chat_id}", response_model=ChatOut)
async def patch_chat(
    chat_id: UUID,
    body: ChatPatch,
    user: CurrentUser,
    session: SessionDep,
) -> ChatOut:
    chat = await _owned_chat(session, user, chat_id)
    if body.title is not None:
        chat.title = body.title
    if body.pinned is not None:
        chat.pinned = body.pinned
    if body.model_id is not None:
        await _assert_model_available(session, body.model_id)
        chat.model_id = body.model_id
    await session.flush()
    return chat_out(
        chat.id,
        chat.title,
        chat.pinned,
        chat.model_id,
        await chat_starter_questions(session, chat),
        chat.created_at,
        None,
    )


@router.delete("/chats/{chat_id}", status_code=204)
async def delete_chat(
    chat_id: UUID,
    user: CurrentUser,
    session: SessionDep,
) -> None:
    chat = await _owned_chat(session, user, chat_id)
    await session.delete(chat)
    await session.flush()


@router.get("/chats/{chat_id}/messages", response_model=list[MessageOut])
async def list_messages(
    chat_id: UUID,
    user: CurrentUser,
    session: SessionDep,
) -> list[MessageOut]:
    await _owned_chat(session, user, chat_id)
    messages = (
        (
            await session.execute(
                select(Message).where(Message.chat_id == chat_id).order_by(Message.created_at)
            )
        )
        .scalars()
        .all()
    )
    citations = (
        await session.execute(
            select(Citation, Chunk, Document, Section)
            .join(Chunk, Citation.chunk_id == Chunk.id, isouter=True)
            .join(Document, Chunk.document_id == Document.id, isouter=True)
            .join(Section, Chunk.section_id == Section.id, isouter=True)
            .where(Citation.message_id.in_([m.id for m in messages]))
            .order_by(Citation.message_id, Citation.n)
        )
    ).all()
    by_message: dict[UUID, list[CitationOut]] = {}
    for citation, chunk, document, section in citations:
        by_message.setdefault(citation.message_id, []).append(
            CitationOut(
                n=citation.n,
                chunk_id=str(chunk.id) if chunk is not None else None,
                document_id=str(document.id) if document is not None else None,
                document_name=document.name if document is not None else None,
                page=chunk.page if chunk is not None else None,
                heading_path=section.heading_path if section is not None else None,
                excerpt=(chunk.text[:240] if chunk is not None else None),
                rerank_score=citation.rerank_score,
                verdict=citation.verdict,
                p_supported=citation.p_supported,
            )
        )
    runs = (
        (await session.execute(select(Run).where(Run.message_id.in_([m.id for m in messages]))))
        .scalars()
        .all()
    )
    metrics_by_message: dict[UUID, dict[str, object] | None] = {
        run.message_id: run.metrics for run in runs
    }
    run_id_by_message: dict[UUID, str] = {run.message_id: str(run.id) for run in runs}
    return [
        message_out(
            m.id,
            m.chat_id,
            m.role,
            m.content,
            m.status,
            m.created_at,
            by_message.get(m.id, []),
            metrics_by_message.get(m.id),
            run_id_by_message.get(m.id),
        )
        for m in messages
    ]


async def _assert_model_available(
    session: AsyncSession, model_id: str
) -> tuple[Model, LlmProvider]:
    row = (
        await session.execute(
            select(Model, LlmProvider)
            .join(LlmProvider, Model.provider_id == LlmProvider.id)
            .where(Model.model_id == model_id, Model.enabled, LlmProvider.enabled)
        )
    ).first()
    if row is None:
        raise ModelNotAvailable("model_not_available", f"Model {model_id} is not enabled")
    model, provider = row
    return model, provider


async def _small_model_litellm(session: AsyncSession, fallback: str) -> str:
    role = (
        await session.execute(select(ModelRole).where(ModelRole.role == "small"))
    ).scalar_one_or_none()
    if role is None:
        return fallback
    row = (
        await session.execute(
            select(LlmProvider.kind)
            .join(Model, Model.provider_id == LlmProvider.id)
            .where(Model.model_id == role.model_id)
        )
    ).scalar_one_or_none()
    if row is None:
        return fallback
    return f"{row}/{role.model_id}"


@router.get("/chats/{chat_id}/documents", response_model=list[DocumentOut])
async def list_chat_documents(
    chat_id: UUID,
    user: CurrentUser,
    session: SessionDep,
) -> list[DocumentOut]:
    chat = await _owned_chat(session, user, chat_id)
    container = await get_chat_collection(session, chat.id)
    if container is None:
        return []
    documents = await list_container_documents(session, container.id)
    return [document_out(document) for document in documents]


@router.post("/chats/{chat_id}/documents", response_model=DocumentUploadOut, status_code=201)
async def upload_chat_document(
    chat_id: UUID,
    file: UploadFile,
    user: CurrentUser,
    session: SessionDep,
) -> DocumentUploadOut:
    chat = await _owned_chat(session, user, chat_id)
    deny_read_only(user)
    container = await get_or_create_chat_collection(session, user, chat)
    filename, data = await read_upload(file)
    document, deduped = await accept_upload(session, container, filename, data)
    return DocumentUploadOut(**document_out(document).model_dump(), deduped=deduped)


@router.post("/chats/{chat_id}/runs", response_model=RunCreateResponse, status_code=201)
async def create_run(
    chat_id: UUID,
    body: RunCreateRequest,
    request: Request,
    user: CurrentUser,
    session: SessionDep,
) -> RunCreateResponse:
    chat = await _owned_chat(session, user, chat_id)
    model_id = body.model_id or chat.model_id
    if model_id is None:
        model_id = await _default_model_id(session)
    model, provider = await _assert_model_available(session, model_id)
    is_first_message = (
        await session.execute(select(Message.id).where(Message.chat_id == chat.id).limit(1))
    ).scalar_one_or_none() is None
    instant_title = None
    if is_first_message and chat.title == DEFAULT_CHAT_TITLE:
        instant_title = collapse_instant_title(body.message)
        chat.title = instant_title

    user_message = Message(chat_id=chat.id, role="user", content=body.message, status="complete")
    assistant_message = Message(chat_id=chat.id, role="assistant", content="", status=None)
    session.add(user_message)
    session.add(assistant_message)
    await session.flush()

    scope = await resolve_scope(session, chat)

    run = Run(
        message_id=assistant_message.id,
        mode=body.mode,
        source=body.source,
        model_id=model_id,
        status="running",
        heartbeat_at=None,
    )
    session.add(run)
    await session.flush()
    runtime = await load_active_runtime(session)
    reservation = await gate_and_reserve(
        session,
        user_id=user.id,
        run_id=run.id,
        mode=body.mode,
        settings=runtime.data,
    )
    run.settings_version = runtime.version
    small_litellm_model = await _small_model_litellm(
        session, f"{provider.kind}/{model_id}"
    )
    await session.commit()

    runner.start_run(
        bus=request.app.state.bus,
        session_factory=request.app.state.session_factory,
        run_id=run.id,
        message_id=assistant_message.id,
        chat_id=chat.id,
        user_id=user.id,
        litellm_model=f"{provider.kind}/{model_id}",
        small_litellm_model=small_litellm_model,
        model_id=model_id,
        context_window=model.context_window or 128_000,
        mode=body.mode,
        source=body.source,
        user_message=body.message,
        run_filters=body.filters,
        collection_ids=scope,
        quota_remaining_5h=reservation.remaining_5h,
        settings_version=runtime.version,
        instant_title=instant_title,
    )
    return RunCreateResponse(run_id=str(run.id), message_id=str(assistant_message.id))